// web/src/api/generated/model/payslipDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PayslipItemDto } from './payslipItemDto';

export interface PayslipDto {
  id: string;
  employeeId: string;
  items: PayslipItemDto[];
  earnings: number;
  deductions: number;
  employeeInsurance: number;
  employerInsurance: number;
  net: number;
}