// web/src/api/generated/model/payslipItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SalaryComponentKind } from './salaryComponentKind';

export interface PayslipItemInput {
  kind: SalaryComponentKind;
  label: string;
  /** @nullable */
  labelAr: string | null;
  amount: number;
  /** @nullable */
  accountId: string | null;
  isInsurable: boolean;
  inEndOfService: boolean;
  /** @nullable */
  componentId?: string | null;
  isBasic?: boolean;
}