// web/src/api/generated/model/salaryComponentKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type SalaryComponentKind = typeof SalaryComponentKind[keyof typeof SalaryComponentKind];


export const SalaryComponentKind = {
  Earning: 'Earning',
  Deduction: 'Deduction',
} as const;